--- a/netty/src/main/java/io/grpc/transport/netty/CancelStreamNettyCommand.java
+++ b/netty/src/main/java/io/grpc/transport/netty/CancelStreamNettyCommand.java
@@ -59,7 +59,7 @@ class CancelStreamNettyCommand extends AbstractNettyCommand {
 
   @Override
   public String toString() {
-    return getClass().getSimpleName() + "(streamId=" + stream.id() + ", flush=" + flush +")";
+    return getClass().getSimpleName() + "(streamId=" + stream.id() + ", flush=" + flush + ")";
   }
 
   @Override