@ejona86 @nmittler 

Update to handle flush reduction change in Netty
Check if we're in the Netty event loop, i.e. a read has occurred and we're using direct executor, when writing so flushes can coalesce

This PR is a work in progress but probably worth discussing now

Partially fixes 
https://github.com/grpc/grpc-java/issues/332
and
https://github.com/grpc/grpc-java/issues/331