@nicolasnoble 2015-05-01
Can one of the admins verify this patch?