@louiscryan 2015-05-09
FYI before and after MaxQpsBenchmakr no's

- We are slower at low channel / stream counts, almost certainly due to delayed flushing and not being saturated.
- As counts go up throughput is significantly better, in particular throughput now scales better with concurrent streams on a single connection to as much as 2x
- Throughput maxes out at 4 connections X 1000 streams per connection and degrades after that but still maintains ~2x margin over old code

BEFORE
Benchmark                                    (channelCount)  (maxConcurrentStreams)   Mode  Cnt      Score      Error  Units
MaxQpsBenchmark.measureUnary:callsPerSecond               1                      10  thrpt   20  17233.925 ±  765.450  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond               1                     100  thrpt   20  17494.865 ± 1069.966  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond               1                    1000  thrpt   20  16806.718 ± 2000.184  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond               2                      10  thrpt   20  26280.909 ± 1800.194  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond               2                     100  thrpt   20  30651.307 ± 2506.293  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond               2                    1000  thrpt   20  30850.247 ± 2760.220  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond               4                      10  thrpt   20  42930.248 ± 1463.644  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond               4                     100  thrpt   20  44527.268 ± 1342.846  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond               4                    1000  thrpt   20  46026.291 ± 2705.484  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond               8                      10  thrpt   20  30151.448 ± 1601.576  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond               8                     100  thrpt   20  35231.488 ±  993.181  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond               8                    1000  thrpt   20  37890.380 ± 4221.883  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond              16                      10  thrpt   20  28447.637 ± 1190.671  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond              16                     100  thrpt   20  30699.913 ±  652.101  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond              16                    1000  thrpt   20  28769.108 ± 3196.432  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond              32                      10  thrpt   20  28440.326 ± 1507.982  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond              32                     100  thrpt   20  29217.955 ± 2022.133  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond              32                    1000  thrpt   20  26925.591 ± 3144.744  ops/s

AFTER--------------------------------------
Benchmark                                    (channelCount)  (maxConcurrentStreams)   Mode  Cnt      Score      Error  Units
MaxQpsBenchmark.measureUnary:callsPerSecond               1                      10  thrpt   20  10733.039 ±  344.622  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond               1                     100  thrpt   20  18730.543 ±  937.766  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond               1                    1000  thrpt   20  39446.294 ± 3000.959  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond               2                      10  thrpt   20  24588.238 ± 1022.536  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond               2                     100  thrpt   20  37055.234 ± 2317.336  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond               2                    1000  thrpt   20  66715.266 ± 4290.143  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond               4                      10  thrpt   20  45346.611 ± 1615.482  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond               4                     100  thrpt   20  63213.694 ± 5365.479  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond               4                    1000  thrpt   20  86546.254 ± 2914.231  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond               8                      10  thrpt   20  74125.868 ± 4187.228  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond               8                     100  thrpt   20  70894.528 ± 2297.936  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond               8                    1000  thrpt   20  58071.587 ± 4204.263  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond              16                      10  thrpt   20  49168.631 ± 2329.883  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond              16                     100  thrpt   20  47208.280 ± 1792.296  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond              16                    1000  thrpt   20  45077.976 ± 4848.871  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond              32                      10  thrpt   20  46096.027 ± 2046.095  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond              32                     100  thrpt   20  43651.986 ± 1615.987  ops/s
MaxQpsBenchmark.measureUnary:callsPerSecond              32                    1000  thrpt   20  38481.980 ± 8374.424  ops/s