@nmittler 2015-05-09
@louiscryan nice results! :)  Any thoughts on how we can improve performance for the few streams case?